♻️  zb: Split internal Node type into separate mod
